refactor: add folders and move files
