+++ src/UI/cards/Cards.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Card from './Card';
+import Card from '../card/Card';
 import style from './Cards.module.scss';
 
 type TCardProps = {
