+++ src/UI/footer/Footer.module.scss
@@ -7,7 +7,7 @@
   display: block;
   width: 80px;
   height: 30px;
-  background-image: url("../assets/svg/rs.svg");
+  background-image: url("../../assets/svg/rs.svg");
   background-position: center;
   background-repeat: no-repeat;
   background-size: contain;
