+++ src/UI/layout/Layout.tsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Outlet } from 'react-router-dom';
 import style from './Layout.module.scss';
-import Header from './Header';
-import Footer from './Footer';
+import Header from '../header/Header';
+import Footer from '../footer/Footer';
 
 class Layout extends Component {
   render() {
